Move Validator to common package, add validation for back slash and add comments.
